backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pymupdf
python-multipart
requests
python-dotenv
transformers
torch

